requirements.txt
streamlit
pillow
pytesseract
requests
pandas
matplotlib
